slightly more breakup and docstrings
