+++ src/data/constants/app.js
@@ -58,7 +58,7 @@ export const messages = StrictDict({
   BulkManagementTab: StrictDict({
     csvUploadLabel: 'Upload Grade CSV',
     heading: 'Use this feature by downloading a CSV for bulk management, overriding grades locally and coming back here to upload',
-    importBtn: 'Import Grades',
+    importBtnText: 'Import Grades',
     successDialog: 'CSV processing. File uploads may take several minutes to complets',
     hints: [
       'Results appear in the table below',
